requirements.txt
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
binance-connector==3.6.0
websocket-client==1.8.0
requests==2.32.3
pytz==2024.1
stable-baselines3==2.3.2
torch>=2.2.0
fastapi==0.111.0
uvicorn==0.30.1
FinRL==0.3.6
pydantic==2.7.2
SQLAlchemy[asyncio]==2.0.30
asyncpg==0.29.0
redis==5.0.5
tenacity==8.3.0
loguru==0.7.2
jinja2==3.1.4
